Refactor: 이미지 영역 컴포넌트 분리 작업
배너 이미지, 소개 이미지 컴포넌트 분리
