+++ app/_utils/generateTimeArray.ts
@@ -1,7 +1,7 @@
 export const generateTimeArray = () => {
   const times = [];
 
-  for (let hour = 0; hour <= 24; hour += 1) {
+  for (let hour = 0; hour <= 23; hour += 1) {
     const formattedHour = String(hour).padStart(2, '0');
     times.push(`${formattedHour}:00`);
   }
